Support URLs as command line argument.
This makes it possible to use smtube with kokovp instead of smplayer.

Formatting

Improvement, to be squashed.

Add missing header in singleinstance.cpp

Add Open URL button to Open menu.

Revert renaming from file to media(Ressource)

To be squashed.

Conditionally start playback of opened media in existing instance.

If something is already currently playing on the existing instance, the
media gets added to the playlist, like before. If nothing is playing,
start playback of the added media.

Add missing QMessageBox header

More cleanup

To be squashed.